+++ src/actions/internal/services/dockerPs.js
@@ -1,5 +1,5 @@
 import { execute } from '../../../utils/execute.js';
 
 export async function dockerPs({ pwd }) {
-  return execute({ pwd, command: `docker-compose ps` });
+  return execute({ pwd, command: `docker compose ps` });
 }