+++ src/actions/internal/tools/fetchMfaCredentials.js
@@ -111,7 +111,7 @@ async function getLoginCredentials({ pwd, serviceName, id, secret, mfa }) {
     );
   }
   const command =
-    `docker-compose exec -T ` +
+    `docker compose exec -T ` +
     `-e AWS_ACCESS_KEY_ID=${id} ` +
     `-e AWS_SECRET_ACCESS_KEY=${secret} ` +
     `${serviceName} ` +