+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+1. Changed from `docker-compose` to `docker compose`
+
 ## [0.13.0][] - 2023-04-13
 
 1. Added `updateMfaProfile` action
@@ -249,7 +251,5 @@
 [0.0.3]: https://github.com/TarSzator/dev-kit/compare/v0.0.2...v0.0.3
 [0.0.2]: https://github.com/TarSzator/dev-kit/compare/v0.0.1...v0.0.2
 [0.0.1]: https://github.com/TarSzator/dev-kit/tree/v0.0.1
-
-
 [Unreleased]: https://github.com/TarSzator/dev-kit/compare/v0.13.0...HEAD
 [0.13.0]: https://github.com/TarSzator/dev-kit/tree/v0.13.0